Revamped
Made the move to make the ORM ready for DI and supporting multiple database configurations.

Database has been renamed to MssqlDatabase implementing an IDatabase;
Constructors support passing a connection string, a configuration file containing the section "HyenaORM" or a configuration section.

Rename of GetAll to LoadAllRecords and GetRecordByPrimaryKey to LoadRecordByPrimaryKey.